@@ -13,20 +13,24 @@ namespace idee5.Common.Data;
 public class DebugValidationReporter : IValidationResultReporter {
     /// <inheritdoc/>
     public void Report(ValidationResult validationResult) {
-        if (validationResult == null)
-            throw new ArgumentNullException(nameof(validationResult));
-
-        Console.WriteLine(String.Format(CultureInfo.InvariantCulture, Resources.MemberNames, validationResult.MemberNames.JoinAsString(",")));
-        Console.WriteLine(String.Format(CultureInfo.InvariantCulture, Resources.Error, validationResult.ErrorMessage));
+#if NETSTANDARD2_0_OR_GREATER
+        if (validationResult == null) throw new ArgumentNullException(nameof(validationResult));
+#else
+        ArgumentNullException.ThrowIfNull(validationResult);
+#endif
+        Console.WriteLine(string.Format(CultureInfo.InvariantCulture, Resources.MemberNames, validationResult.MemberNames.JoinAsString(",")));
+        Console.WriteLine(string.Format(CultureInfo.InvariantCulture, Resources.Error, validationResult.ErrorMessage));
     }
 
     /// <inheritdoc/>
     public Task ReportAsync(ValidationResult validationResult, CancellationToken cancellationToken = default) {
-        if (validationResult == null)
-            throw new ArgumentNullException(nameof(validationResult));
-
-        Console.WriteLine(String.Format(CultureInfo.InvariantCulture, Resources.MemberNames, validationResult.MemberNames.JoinAsString(",")));
-        Console.WriteLine(String.Format(CultureInfo.InvariantCulture, Resources.Error, validationResult.ErrorMessage));
+#if NETSTANDARD2_0_OR_GREATER
+        if (validationResult == null) throw new ArgumentNullException(nameof(validationResult));
+#else
+        ArgumentNullException.ThrowIfNull(validationResult);
+#endif
+        Console.WriteLine(string.Format(CultureInfo.InvariantCulture, Resources.MemberNames, validationResult.MemberNames.JoinAsString(",")));
+        Console.WriteLine(string.Format(CultureInfo.InvariantCulture, Resources.Error, validationResult.ErrorMessage));
         return Task.CompletedTask;
     }
 }
