@@ -12,13 +12,15 @@ public static class TypeExtensions {
     /// <param name="type"></param>
     /// <returns>The <see cref="IEnumerable{T}"/> with all constant values.</returns>
     /// <exception cref="ArgumentNullException"><paramref name="type"/> is <c>null</c>.</exception>
-    public static IEnumerable<T> GetAllPublicConstantValues<T>(this Type type) {
-        if (type == null)
-            throw new ArgumentNullException(nameof(type));
-
+    public static IEnumerable<T?> GetAllPublicConstantValues<T>(this Type type) {
+#if NETSTANDARD2_0_OR_GREATER
+        if (type == null) throw new ArgumentNullException(nameof(type));
+#else
+        ArgumentNullException.ThrowIfNull(type);
+#endif
         return type.GetFields(BindingFlags.Public | BindingFlags.Static | BindingFlags.FlattenHierarchy)
             .Where(fi => fi.IsLiteral && !fi.IsInitOnly && fi.FieldType == typeof(T))
-            .Select(x => (T)x.GetValue(null));
+            .Select(x => (T?)x.GetValue(null));
     }
 
     /// <summary>
@@ -28,10 +30,13 @@ public static class TypeExtensions {
     /// <returns>True if it is a CLR type.</returns>
     /// <exception cref="ArgumentNullException"><paramref name="type"/> is <c>null</c>.</exception>
     public static bool IsClrType(this Type type) {
-        if (type == null)
-            throw new ArgumentNullException(nameof(type));
+#if NETSTANDARD2_0_OR_GREATER
+        if (type == null) throw new ArgumentNullException(nameof(type));
+#else
+        ArgumentNullException.ThrowIfNull(type);
+#endif
 
-        string fullname = type.Assembly.FullName;
+        string fullname = type.Assembly.FullName ?? "";
         // check new (net core), then old (full framework) .net
         // Assuming more usage of .net core, we check that first.
         return fullname.StartsWith("System.Private.CoreLib", StringComparison.InvariantCulture)
