@@ -2,17 +2,15 @@ using System.Collections.Generic;
 using System.Globalization;
 
 namespace idee5.Common;
+
 /// <summary>
 /// Parameters for querying the available countries.
 /// </summary>
-public class CountryQuery : IQuery<IDictionary<string, CultureInfo>> {
-    /// <summary>
-    /// Partial country name.
-    /// </summary>
-    public string NameFilter { get; set; }
-
+/// <param name="NameFilter"> Partial country name</param>
+public record CountryQuery(string NameFilter) : IQuery<IDictionary<string, CultureInfo>> {
     /// <summary>
     /// Initializes a new instance of the <see cref="CountryQuery"/> class.
     /// </summary>
-    public CountryQuery() { NameFilter = ""; }
+    public CountryQuery() : this("") {
+    }
 }
