@@ -2,30 +2,18 @@
 using System.Globalization;
 
 namespace idee5.Common;
+
+
 /// <summary>
 /// Query the available languages with filtering.
 /// </summary>
-public class LanguageQuery : IQuery<IDictionary<string, CultureInfo>> {
-    /// <summary>
-    /// Text to search for in the language name.
-    /// </summary>
-    public string LanguageFilter { get; set; }
-
+/// <param name="LanguageFilter"> Text to search for in the language name. </param>
+/// <param name="IsCaseSensitiveQuery"> Do a case sensitive filtering or not. </param>
+/// <param name="TypeFilter"> <see cref="CultureTypes"/> to be queried. </param>
+public record LanguageQuery(string LanguageFilter, bool IsCaseSensitiveQuery, CultureTypes TypeFilter) : IQuery<IDictionary<string, CultureInfo>> {
     /// <summary>
     /// Creates a new language query.
     /// </summary>
-    public LanguageQuery() {
-        LanguageFilter = "";
-        IsCaseSensitiveQuery = false;
+    public LanguageQuery() : this("", false, default) {
     }
-
-    /// <summary>
-    /// Do a case sensitive filtering or not.
-    /// </summary>
-    public bool IsCaseSensitiveQuery { get; set; }
-
-    /// <summary>
-    /// <see cref="CultureTypes"/> to be queried.
-    /// </summary>
-    public CultureTypes TypeFilter { get; set; }
 }
